EDU-348 refactor: 사전 사후 문제 데이터 저장
